Add HTML pages for Cafe Quest project including homepage, login page, my events page, and student authentication page with styles and basic functionality
