+++ project/backend-api/src/main/java/com/example/CafeQuestBackendAPI/controller/CoffeeShopController.java
@@ -1,13 +1,21 @@
 package com.example.CafeQuestBackendAPI.controller;
 
-import com.example.CafeQuestBackendAPI.entity.CoffeeShop;
-import com.example.CafeQuestBackendAPI.service.CoffeeShopService;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import com.example.CafeQuestApiApplication.entity.CoffeeShop;
+import com.example.CafeQuestApiApplication.service.CoffeeShopService;
 
 @RestController
 @RequestMapping("/api/coffeeshops")
