+++ project/backend-api/src/main/java/com/example/CafeQuestBackendAPI/controller/ProfileController.java
@@ -1,7 +1,7 @@
-package com.example.CafeQuestBackendAPI.controller;
+package com.example.CafeQuestApiApplication.controller;
 
-import com.example.CafeQuestBackendAPI.entity.Profile;
-import com.example.CafeQuestBackendAPI.service.ProfileService;
+import com.example.CafeQuestApiApplication.entity.Profile;
+import com.example.CafeQuestApiApplication.service.ProfileService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/profiles")
+@CrossOrigin(origins = "*")
 public class ProfileController {
     
     @Autowired
