+++ project/backend-api/src/main/java/com/example/CafeQuestBackendAPI/controller/ReviewController.java
@@ -1,7 +1,7 @@
-package com.example.CafeQuestBackendAPI.controller;
+package com.example.CafeQuestApiApplication.controller;
 
-import com.example.CafeQuestBackendAPI.entity.Review;
-import com.example.CafeQuestBackendAPI.service.ReviewService;
+import com.example.CafeQuestApiApplication.entity.Review;
+import com.example.CafeQuestApiApplication.service.ReviewService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,6 +11,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/api/reviews")
+@CrossOrigin(origins = "*")
 public class ReviewController {
     
     @Autowired
