+++ project/backend-api/src/main/java/com/example/CafeQuestBackendAPI/controller/UserOrderController.java
@@ -1,7 +1,7 @@
-package com.example.CafeQuestBackendAPI.controller;
+package com.example.CafeQuestApiApplication.controller;
 
-import com.example.CafeQuestBackendAPI.entity.UserOrder;
-import com.example.CafeQuestBackendAPI.service.UserOrderService;
+import com.example.CafeQuestApiApplication.entity.UserOrder;
+import com.example.CafeQuestApiApplication.service.UserOrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,6 +11,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/api/orders")
+@CrossOrigin(origins = "*")
 public class UserOrderController {
     
     @Autowired
