+++ project/backend-api/src/main/java/com/example/CafeQuestBackendAPI/entity/Profile.java
@@ -1,4 +1,4 @@
-package com.example.CafeQuestBackendAPI.entity;
+package com.example.CafeQuestApiApplication.entity;
 
 import jakarta.persistence.*;
