+++ project/backend-api/src/main/java/com/example/CafeQuestBackendAPI/entity/User.java
@@ -1,4 +1,4 @@
-package com.example.CafeQuestBackendAPI.entity;
+package com.example.CafeQuestApiApplication.entity;
 
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
