+++ project/backend-api/src/main/java/com/example/CafeQuestBackendAPI/repository copy/ProfileRepository.java
@@ -1,0 +1,16 @@
+package com.example.CafeQuestApiApplication.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+
+import com.example.CafeQuestApiApplication.entity.Profile;
+
+@Repository
+public interface ProfileRepository extends JpaRepository<Profile, Long> {
+
+    @Query("SELECT p FROM Profile p WHERE p.user.userId = ?1")
+    Optional<Profile> findByUser_UserId(Long userId);
+}
