+++ project/backend-api/src/main/java/com/example/CafeQuestBackendAPI/repository copy/ReviewRepository.java
@@ -1,4 +1,4 @@
-package com.example.CafeQuestBackendAPI.repository;
+package com.example.CafeQuestApiApplication.repository;
 
 import java.util.List;
 
@@ -6,18 +6,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import com.example.CafeQuestBackendAPI.entity.Review;
+import com.example.CafeQuestApiApplication.entity.Review;
 
 @Repository
 public interface ReviewRepository extends JpaRepository<Review, Long> {
     
-    @Query("SELECT r FROM Review r WHERE r.user.fullName LIKE %?1%")
-    List<Review> getReviewsByUserName(String userName);
-
     @Query("SELECT r FROM Review r WHERE r.coffeeShop.shopId = ?1")
     List<Review> findByCoffeeShop_ShopId(Long shopId);
 
     @Query("SELECT r FROM Review r WHERE r.user.userId = ?1")
     List<Review> findByUser_UserId(Long userId);
-
 }
