+++ project/backend-api/src/main/java/com/example/CafeQuestBackendAPI/repository copy/UserRepository.java
@@ -1,11 +1,11 @@
-package com.example.CafeQuestBackendAPI.repository;
+package com.example.CafeQuestApiApplication.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.CafeQuestBackendAPI.entity.User;
+import com.example.CafeQuestApiApplication.entity.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
