+++ project/backend-api/src/main/resources/static/project/homepage.html
@@ -1,0 +1,143 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Cafe Quest API</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            max-width: 800px;
+            margin: 50px auto;
+            padding: 20px;
+            background-color: #f4efe7;
+        }
+        h1 {
+            color: #3e2723;
+        }
+        .endpoint {
+            background: white;
+            padding: 15px;
+            margin: 10px 0;
+            border-radius: 8px;
+            border: 1px solid #3e2723;
+        }
+        .method {
+            display: inline-block;
+            padding: 5px 10px;
+            border-radius: 4px;
+            font-weight: bold;
+            margin-right: 10px;
+        }
+        .get { background-color: #61affe; color: white; }
+        .post { background-color: #49cc90; color: white; }
+        .put { background-color: #fca130; color: white; }
+        .delete { background-color: #f93e3e; color: white; }
+        a {
+            color: #3e2723;
+            text-decoration: none;
+        }
+        a:hover {
+            text-decoration: underline;
+        }
+        .link-section {
+            margin: 30px 0;
+            padding: 20px;
+            background: white;
+            border-radius: 8px;
+            border: 1px solid #3e2723;
+        }
+    </style>
+</head>
+<body>
+    <h1>☕ Cafe Quest API</h1>
+    <p>Welcome to the Cafe Quest</p>
+    
+    <div class="link-section">
+        <h2>Frontend Pages</h2>
+        <ul>
+            <li><a href="/project/loginpage.html">Login Page</a></li>
+            <li><a href="/project/homepage.html">Home Page</a></li>
+            <li><a href="/project/myevents.html">My Events</a></li>
+        </ul>
+    </div>
+
+    <h2>API Endpoints</h2>
+    
+    <h3>Coffee Shops</h3>
+    <div class="endpoint">
+        <span class="method get">GET</span>
+        <a href="/api/coffeeshops">/api/coffeeshops</a> - Get all coffee shops
+    </div>
+    <div class="endpoint">
+        <span class="method get">GET</span>
+        /api/coffeeshops/{id} - Get coffee shop by ID
+    </div>
+    <div class="endpoint">
+        <span class="method post">POST</span>
+        /api/coffeeshops - Create new coffee shop
+    </div>
+    <div class="endpoint">
+        <span class="method put">PUT</span>
+        /api/coffeeshops/{id} - Update coffee shop
+    </div>
+    <div class="endpoint">
+        <span class="method delete">DELETE</span>
+        /api/coffeeshops/{id} - Delete coffee shop
+    </div>
+
+    <h3>Users</h3>
+    <div class="endpoint">
+        <span class="method post">POST</span>
+        /api/users/signup - Sign up new user
+    </div>
+    <div class="endpoint">
+        <span class="method post">POST</span>
+        /api/users/login - Login
+    </div>
+    <div class="endpoint">
+        <span class="method get">GET</span>
+        <a href="/api/users">/api/users</a> - Get all users
+    </div>
+
+    <h3>Events</h3>
+    <div class="endpoint">
+        <span class="method get">GET</span>
+        <a href="/api/events">/api/events</a> - Get all events
+    </div>
+    <div class="endpoint">
+        <span class="method post">POST</span>
+        /api/events/user/{userId}/shop/{shopId} - Create event
+    </div>
+
+    <h3>Reviews</h3>
+    <div class="endpoint">
+        <span class="method post">POST</span>
+        /api/reviews/user/{userId}/shop/{shopId} - Create review
+    </div>
+    <div class="endpoint">
+        <span class="method get">GET</span>
+        /api/reviews/shop/{shopId} - Get reviews for a shop
+    </div>
+
+    <h3>Orders</h3>
+    <div class="endpoint">
+        <span class="method post">POST</span>
+        /api/orders/user/{userId} - Create order
+    </div>
+    <div class="endpoint">
+        <span class="method get">GET</span>
+        /api/orders/user/{userId} - Get user orders
+    </div>
+
+    <h3>Profiles</h3>
+    <div class="endpoint">
+        <span class="method get">GET</span>
+        /api/profiles/user/{userId} - Get user profile
+    </div>
+    <div class="endpoint">
+        <span class="method put">PUT</span>
+        /api/profiles/user/{userId} - Update user profile
+    </div>
+</body>
+</html>
