+++ project/backend-api/src/main/resources/static/project/myevents.html
@@ -1,0 +1,113 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Cafe Quest - My Events</title>
+  <style>
+    body {
+      font-family: Arial, sans-serif;
+      margin: 0;
+      background-color: #fdf6f0;
+    }
+    header {
+      display: flex;
+      align-items: center;
+      padding: 15px;
+      background-color: #d7ccc8;
+    }
+    header .logo-space {
+      width: 50px;
+      height: 50px;
+      background-color: #efebe9;
+      border-radius: 8px;
+      margin-right: 10px;
+    }
+    header h1 {
+      font-size: 22px;
+      color: #5d4037;
+    }
+    main {
+      padding: 20px;
+    }
+    .event-card {
+      background-color: #fff;
+      border-radius: 12px;
+      padding: 15px;
+      margin-bottom: 15px;
+      box-shadow: 0 4px 10px rgba(0,0,0,0.1);
+    }
+    .event-card h2 {
+      margin: 0;
+      color: #5d4037;
+    }
+    .event-card p {
+      margin: 5px 0;
+      color: #6d4c41;
+    }
+    nav {
+      position: fixed;
+      bottom: 0;
+      left: 0;
+      right: 0;
+      display: flex;
+      justify-content: space-around;
+      background-color: #d7ccc8;
+      padding: 10px 0;
+      border-top: 2px solid #a1887f;
+    }
+    nav div {
+      text-align: center;
+      color: #5d4037;
+      font-size: 14px;
+    }
+    nav img {
+      width: 28px;
+      height: 28px;
+      border-radius: 6px;
+      display: block;
+      margin: 0 auto 5px;
+    }
+  </style>
+</head>
+<body>
+  <header>
+    <div class="logo-space"></div>
+    <h1>My Events</h1>
+  </header>
+
+  <main>
+    <div class="event-card">
+      <h2>CSC-330 Study Group</h2>
+      <p>📍 Jackson Library, Room 204</p>
+      <p>🗓️ Sept 25, 6:00 PM</p>
+    </div>
+    <div class="event-card">
+      <h2>Poetry Night</h2>
+      <p>📍 Tate Street Coffee</p>
+      <p>🗓️ Sept 28, 7:30 PM</p>
+    </div>
+  </main>
+
+  <nav>
+    <div>
+      <a href="home.html">
+        <img src="https://i.pinimg.com/1200x/c3/41/6e/c3416efade9f62b399498e657113ddf7.jpg" alt="Home">
+      </a>
+      Home
+    </div>
+    <div>
+      <a href="events.html">
+        <img src="https://i.pinimg.com/736x/bb/86/e7/bb86e710bc7167e62f7982e4d04de5e2.jpg" alt="My Events">
+      </a>
+      My Events
+    </div>
+    <div>
+      <a href="profile.html">
+        <img src="https://i.pinimg.com/736x/71/4a/cf/714acfb1b90df376dfa9906c1e316c67.jpg" alt="Me">
+      </a>
+      Me
+    </div>
+  </nav>
+</body>
+</html>
