+++ project/backend-api/src/main/resources/static/project/studentauth.html
@@ -1,0 +1,240 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Cafe Quest</title>
+  <style>
+    body {
+      margin: 0;
+      font-family: Arial, sans-serif;
+      background-color: #f5f0e6;
+      color: #3e2723;
+    }
+
+    .page {
+      display: none;
+      height: 100vh;
+    }
+
+    /* Login */
+    #auth-page {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      background: linear-gradient(135deg, #6d4c41, #a1887f);
+    }
+
+    .auth-box {
+      background: #fff;
+      padding: 30px;
+      border-radius: 12px;
+      box-shadow: 0 6px 18px rgba(0,0,0,0.25);
+      width: 320px;
+      text-align: center;
+    }
+
+    .auth-box h2 {
+      margin-bottom: 20px;
+    }
+
+    .auth-box input {
+      width: 100%;
+      padding: 12px;
+      margin: 10px 0;
+      border: 1px solid #ccc;
+      border-radius: 8px;
+    }
+
+    .auth-box button {
+      width: 100%;
+      padding: 12px;
+      background-color: #6d4c41;
+      color: #fff;
+      border: none;
+      border-radius: 8px;
+      font-size: 16px;
+      cursor: pointer;
+    }
+
+    .auth-box button:hover {
+      background-color: #4e342e;
+    }
+
+    /* Home page */
+    #home-page {
+      display: flex;
+      flex-direction: column;
+      height: 100vh;
+    }
+
+    header {
+      display: flex;
+      align-items: center;
+      padding: 15px;
+      background-color: #d7ccc8;
+    }
+
+    header .logo-space {
+      width: 50px;
+      height: 50px;
+      background: #bbb;
+      border-radius: 8px;
+      margin-right: 15px;
+    }
+
+    header h1 {
+      font-size: 20px;
+      margin: 0;
+    }
+
+    .content {
+      flex: 1;
+      overflow-y: auto;
+      padding: 15px;
+    }
+
+    .shop {
+      display: flex;
+      flex-direction: row;
+      background: #fff;
+      border-radius: 12px;
+      box-shadow: 0 2px 6px rgba(0,0,0,0.15);
+      margin-bottom: 20px;
+      overflow: hidden;
+    }
+
+    .shop img {
+      width: 60%;
+      background: #d7ccc8;
+      object-fit: cover;
+    }
+
+    .shop-info {
+      width: 40%;
+      padding: 10px;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+    }
+
+    .shop-info div {
+      margin: 6px 0;
+      padding: 6px;
+      background: #efebe9;
+      border-radius: 6px;
+      text-align: center;
+    }
+
+    /* Bottom nav */
+    nav {
+      display: flex;
+      justify-content: space-around;
+      align-items: center;
+      background-color: #a1887f;
+      padding: 10px 0;
+    }
+
+    nav div {
+      text-align: center;
+      color: #fff;
+      font-size: 14px;
+    }
+
+    nav i {
+      display: block;
+      font-size: 22px;
+      margin-bottom: 4px;
+    }
+  </style>
+</head>
+<body>
+
+  <!-- Login Page -->
+  <div id="auth-page" class="page" style="display: flex;">
+    <div class="auth-box">
+      <h2>Login</h2>
+      <input type="email" id="email" placeholder="Student Email" required>
+      <input type="password" id="password" placeholder="Password" required>
+      <button onclick="validateLogin()">Login</button>
+    </div>
+  </div>
+
+  <!-- Home Page -->
+  <div id="home-page" class="page">
+    <header>
+      <div class="logo-space"></div>
+      <h1>Cafe Quest</h1>
+    </header>
+
+    <div class="content">
+      <div class="shop">
+        <img src="" alt="Coffee Shop Image">
+        <div class="shop-info">
+          <div>123 Tate St</div>
+          <div>Comfort: Cozy chairs</div>
+          <div>Rating ☕☕☕</div>
+        </div>
+      </div>
+
+      <div class="shop">
+        <img src="" alt="Coffee Shop Image">
+        <div class="shop-info">
+          <div>456 Spring Garden St</div>
+          <div>Comfort: Spacious tables</div>
+          <div>Rating ☕☕</div>
+        </div>
+      </div>
+    </div>
+
+    <nav>
+        <div>
+            <!-- <a href="https://www.uncg.edu"> -->
+                <img src="https://i.pinimg.com/1200x/c3/41/6e/c3416efade9f62b399498e657113ddf7.jpg" 
+                 alt="Home" 
+                style="width:30px; height:30px; object-fit:cover; border-radius:6px;">
+            </a>
+            Home
+        </div>
+    <nav>
+        <div>
+            <!-- <a href="https://www.uncg.edu"> -->
+                <img src="https://i.pinimg.com/736x/bb/86/e7/bb86e710bc7167e62f7982e4d04de5e2.jpg" 
+                 alt="My Events" 
+                style="width:30px; height:30px; object-fit:cover; border-radius:6px;">
+            </a>
+            My Events
+        </div>
+    <nav>
+        <div>
+            <!--<a href="https://www.uncg.edu"> -->
+                <img src="https://i.pinimg.com/736x/71/4a/cf/714acfb1b90df376dfa9906c1e316c67.jpg" 
+                 alt="Me" 
+                style="width:30px; height:30px; object-fit:cover; border-radius:6px;">
+            </a>
+            Me
+        </div>
+
+  <script>
+    function validateLogin() {
+      const email = document.getElementById("email").value;
+      const password = document.getElementById("password").value;
+
+      if (!email.endsWith("@uncg.edu")) {
+        alert("Please use your UNCG student email (must end with @uncg.edu).");
+        return;
+      }
+
+      if (password.trim() === "") {
+        alert("Please enter your password.");
+        return;
+      }
+
+      // Show home page if valid
+      document.getElementById("auth-page").style.display = "none";
+      document.getElementById("home-page").style.display = "flex";
+      document.getElementById("home-page").style.flexDirection = "column";
+    }
+  </script>
+</body>
+</html>
